requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=./requirements/requirements.txt ./requirements/requirements.in
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.13.0
    # via
    #   -r ./requirements/requirements.in
    #   cdk-hypderledger-fabric-network
cattrs==1.10.0
    # via jsii
cdk-hypderledger-fabric-network==0.0.7
    # via -r ./requirements/requirements.in
constructs==10.0.69
    # via
    #   -r ./requirements/requirements.in
    #   aws-cdk-lib
    #   cdk-hypderledger-fabric-network
jsii==1.54.0
    # via
    #   aws-cdk-lib
    #   cdk-hypderledger-fabric-network
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   cdk-hypderledger-fabric-network
    #   constructs
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typing-extensions==4.1.1
    # via jsii
